custom fixture and reorganize tests
